flaskProject/requirements.txt
mysql
mysql-connector-python
flask-restful